♻️refactor: Taglnput js->tsx 변경
